Refactored Security
